Create preprocess.js
