+++ src/distance.js
@@ -1,35 +1,8 @@
 var LinearInterpolation = require('./LinearInterpolation.js');
+var Preprocess = require('./preprocess.js');
 var DTW = require('./dtw.js');
 
 var Distance = {};
-
-
-function normalize(value, min, max) {
-    return (value - min) / (max - min);
-}
-
-function normalizeArray(array, min, max) {
-    var narray = [];
-    for (var i = 0; i < array.length; i++) {
-	var nv = normalize(array[i], min, max);
-	narray.push(nv);
-    }
-    return narray;
-}
-
-function normalizeXYZArray(array, min, max) {
-    var narray = [];
-    for (var i = 0; i < array.length; i++) {
-	var p = array[i];
-	var np = {
-	    x: normalize(p.x, min, max),
-	    y: normalize(p.y, min, max),
-	    z: normalize(p.z, min, max)
-	};
-	narray.push(np);
-    }
-    return narray;
-}
 
 // 1次元の点同士のユークリッド距離を求める関数
 // 1次元の場合のユークリッド距離は差の絶対値に等しい
@@ -39,43 +12,15 @@ function distance1D(p1, p2) {
 };
 
 
-// 3次元時系列データの変移を求める関数
-// 1次元時系列データを返す
-function changeOfDistance(ts) {
-    var n = ts.length - 1;
-    var d = [];
-    for (var i = 0; i < n; i++) {
-	var x = Math.pow(ts[i+1].x - ts[i].x, 2);
-	var y = Math.pow(ts[i+1].y - ts[i].y, 2);
-	var z = Math.pow(ts[i+1].z - ts[i].z, 2);
-	d.push(Math.sqrt(x + y + z));
-    }
-    return d;
-}
-
-// 1次元時系列データをその最小値から最大値の間の値に正規化する
-function temporalNormalize(ts) {
-    var max = Math.max.apply(null, ts);
-    var min = Math.min.apply(null, ts);
-    
-    if (min == max) {
-	// create a '0.5' filled array
-	var ret = Array.apply(null, Array(ts.length)).map(Number.prototype.valueOf, 0.5);
-	return ret;
-    }
-
-    return normalizeArray(ts, min, max);
-}
-
 // 時間的類似度を求めるための前処理
 function temporalPreprocess(ts) {
     // linear interpolation?
 
     // change of distance
-    var ts_cod = changeOfDistance(ts);
+    var ts_cod = Preprocess.changeOfDistance(ts);
 
     // normalize
-    var ts_cod_n = temporalNormalize(ts_cod);
+    var ts_cod_n = Preprocess.temporalNormalize(ts_cod);
 
     return ts_cod_n;    
 }
@@ -90,24 +35,6 @@ Distance.temporalDistance = function(ts1, ts2) {
     return d;
 };
 
-
-// 3次元時系列データをその最小値から最大値の間に正規化する
-function spatialNormalize(ts) {
-    var d = ts.map(function(d) { return [d.x, d.y, d.z]; });
-    var ary = Array.prototype.concat.apply([], d);
-    var min = Math.min.apply(null, ary);
-    var max = Math.max.apply(null, ary);
-
-    if (min == max) {
-	// create a '0.5' filled xyz array
-	var ret = Array.apply(null, Array(ts.length)).map(function() {
-	    return {x: 0.5, y: 0.5, z: 0.5};
-	});
-	return ret;
-    }
-
-    return normalizeXYZArray(ts, min, max);
-}
 
 // 空間的類似度を求めるための前処理
 function spatialPreprocess(ts) {
@@ -128,7 +55,7 @@ function spatialPreprocess(ts) {
     var ts_li_z0 = ignoreZAxis(ts_li);
     
     // normalize
-    var ts_li_z0_n = spatialNormalize(ts_li_z0);
+    var ts_li_z0_n = Preprocess.spatialNormalize(ts_li_z0);
 
     // get x-axis
     var ts_x = ts_li_z0_n.map(function(e) { return e.x; });
